New feature - check if sum of input's digits is divisible by 8
If sum of all input integer's digits is divisible by 8, `Inf` is appended to
the output.

Changes in the class `FooBarQix`:

- property `$_ouput` was changed from private to protected to make it
  accessible in subclasses and renamed to `$output`;
- constructor method was changed from private to protected to make it
  accessible in subclasses;
- new method `inputDigitsAsArray` was added to get an array that contains input
  integer's digits;
- method `_checkOccurrences` was updated to use the new method
  `inputDigitsAsArray` to get the necessary value.

The class `InfQixFoo` was updated to add the functionality to check if the sum
of input integer's digits is divisible by 8 (Inf) to add `Inf` to the output in
such cases.

`tests/InfQixFooTest.php` was updated to provide values in some tests that do
not clash with the new functionality.