tests: Add test coverage support
Configured using xdebug driver.

Currently, one line is not covered. That is the default case in a switch
statement in BaseNumberMapperService::mapNumber(), but since it is a
protected method, it is not tested directly with unit tests.

Time: 00:00.177, Memory: 18.00 MB

OK (31 tests, 100 assertions)

Code Coverage Report:
  2021-05-17 10:09:34

 Summary:
  Classes: 66.67% (2/3)
  Methods: 88.89% (8/9)
  Lines:   97.73% (43/44)

FooBar\Services\BaseNumberMapperService
  Methods:  83.33% ( 5/ 6)   Lines:  97.44% ( 38/ 39)
FooBar\Services\FooBarQixService
  Methods: 100.00% ( 2/ 2)   Lines: 100.00% (  2/  2)
FooBar\Services\InfQixFooService
  Methods: 100.00% ( 1/ 1)   Lines: 100.00% (  3/  3)
Remote CodeCoverage reports are not printed to console

Signed-off-by: Reinis Danne <rei4dan@gmail.com>